+++ crates/l2/proposer/l1_commiter.rs
@@ -1,0 +1,441 @@
+use std::{collections::HashMap, time::Duration};
+
+use crate::{
+    proposer::state_diff::{AccountStateDiff, DepositLog, WithdrawalLog},
+    utils::{
+        config::{eth::EthConfig, proposer::ProposerConfig},
+        eth_client::{transaction::blob_from_bytes, EthClient},
+        merkle_tree::merkelize,
+    },
+};
+use bytes::Bytes;
+use c_kzg::{Bytes48, KzgSettings};
+use ethereum_rust_blockchain::constants::TX_GAS_COST;
+use ethereum_rust_core::{
+    types::{
+        BlobsBundle, Block, EIP1559Transaction, EIP4844Transaction, GenericTransaction,
+        PrivilegedL2Transaction, PrivilegedTxType, Transaction, TxKind, BYTES_PER_BLOB,
+    },
+    Address, H256, U256,
+};
+use ethereum_rust_rpc::types::transaction::WrappedEIP4844Transaction;
+use ethereum_rust_storage::Store;
+use ethereum_rust_vm::{evm_state, execute_block, get_state_transitions};
+use keccak_hash::keccak;
+use libsecp256k1::SecretKey;
+use sha2::{Digest, Sha256};
+use tokio::time::sleep;
+use tracing::{error, info};
+
+use super::{errors::ProposerError, state_diff::StateDiff};
+use crate::utils::eth_client::{errors::EthClientError, eth_sender::Overrides};
+
+const COMMIT_FUNCTION_SELECTOR: [u8; 4] = [132, 97, 12, 179];
+
+pub struct Commiter {
+    eth_client: EthClient,
+    on_chain_proposer_address: Address,
+    store: Store,
+    l1_address: Address,
+    l1_private_key: SecretKey,
+    kzg_settings: &'static KzgSettings,
+}
+
+pub async fn start_l1_commiter(store: Store) {
+    let eth_config = EthConfig::from_env().expect("EthConfig::from_env()");
+    let proposer_config = ProposerConfig::from_env().expect("ProposerConfig::from_env");
+    let commiter = Commiter::new_from_config(&proposer_config, eth_config, store).unwrap();
+    let _ = commiter.start().await;
+}
+
+impl Commiter {
+    pub fn new_from_config(
+        proposer_config: &ProposerConfig,
+        eth_config: EthConfig,
+        store: Store,
+    ) -> Result<Self, ProposerError> {
+        Ok(Self {
+            eth_client: EthClient::new(&eth_config.rpc_url),
+            on_chain_proposer_address: proposer_config.on_chain_proposer_address,
+            store,
+            l1_address: proposer_config.l1_address,
+            l1_private_key: proposer_config.l1_private_key,
+            kzg_settings: c_kzg::ethereum_kzg_settings(),
+        })
+    }
+
+    pub async fn start(&self) -> Result<(), ProposerError> {
+        loop {
+            let last_commited_block = get_last_commited_block(
+                &self.eth_client,
+                self.on_chain_proposer_address,
+                Overrides::default(),
+            )
+            .await?;
+
+            let last_commited_block = last_commited_block
+                .strip_prefix("0x")
+                .expect("Couldn't strip prefix from last_commited_block.");
+
+            let last_commited_block = U256::from_str_radix(last_commited_block, 16)
+                .unwrap()
+                .as_u64();
+
+            let block_number_to_fetch = if last_commited_block == u64::MAX {
+                0
+            } else {
+                last_commited_block + 1
+            };
+
+            if let Some(block_to_commit_body) =
+                self.store.get_block_body(block_number_to_fetch).unwrap()
+            {
+                let block_to_commit_header = self
+                    .store
+                    .get_block_header(block_number_to_fetch)
+                    .unwrap()
+                    .unwrap();
+
+                let block_to_commit = Block::new(block_to_commit_header, block_to_commit_body);
+
+                let withdrawals = self.get_block_withdrawals(&block_to_commit)?;
+                let deposits = self.get_block_deposits(&block_to_commit)?;
+
+                let withdrawal_logs_merkle_root = self.get_withdrawals_merkle_root(
+                    withdrawals.iter().map(|(hash, _tx)| *hash).collect(),
+                );
+                let deposit_logs_hash = self.get_deposit_hash(
+                    deposits
+                        .iter()
+                        .filter_map(|tx| tx.get_deposit_hash())
+                        .collect(),
+                );
+
+                let state_diff = self.prepare_state_diff(
+                    &block_to_commit,
+                    self.store.clone(),
+                    withdrawals,
+                    deposits,
+                )?;
+
+                let (blob_commitment, blob_proof) =
+                    self.prepare_blob_commitment(state_diff.clone())?;
+
+                let head_block_hash = block_to_commit.hash();
+                match self
+                    .send_commitment(
+                        block_to_commit.header.number,
+                        withdrawal_logs_merkle_root,
+                        deposit_logs_hash,
+                        blob_commitment,
+                        blob_proof,
+                        state_diff.encode()?,
+                    )
+                    .await
+                {
+                    Ok(commit_tx_hash) => {
+                        info!(
+                    "Sent commitment to block {head_block_hash:#x}, with transaction hash {commit_tx_hash:#x}"
+                );
+                    }
+                    Err(error) => {
+                        error!("Failed to send commitment to block {head_block_hash:#x}. Manual intervention required: {error}");
+                        panic!("Failed to send commitment to block {head_block_hash:#x}. Manual intervention required: {error}");
+                    }
+                }
+            }
+
+            sleep(Duration::from_secs(1)).await;
+        }
+    }
+
+    pub fn get_block_withdrawals(
+        &self,
+        block: &Block,
+    ) -> Result<Vec<(H256, PrivilegedL2Transaction)>, ProposerError> {
+        let withdrawals = block
+            .body
+            .transactions
+            .iter()
+            .filter_map(|tx| match tx {
+                Transaction::PrivilegedL2Transaction(priv_tx)
+                    if priv_tx.tx_type == PrivilegedTxType::Withdrawal =>
+                {
+                    Some((tx.compute_hash(), priv_tx.clone()))
+                }
+                _ => None,
+            })
+            .collect();
+
+        Ok(withdrawals)
+    }
+
+    pub fn get_withdrawals_merkle_root(&self, withdrawals_hashes: Vec<H256>) -> H256 {
+        if !withdrawals_hashes.is_empty() {
+            merkelize(withdrawals_hashes)
+        } else {
+            H256::zero()
+        }
+    }
+
+    pub fn get_block_deposits(
+        &self,
+        block: &Block,
+    ) -> Result<Vec<PrivilegedL2Transaction>, ProposerError> {
+        let deposits = block
+            .body
+            .transactions
+            .iter()
+            .filter_map(|tx| match tx {
+                Transaction::PrivilegedL2Transaction(tx)
+                    if tx.tx_type == PrivilegedTxType::Deposit =>
+                {
+                    Some(tx.clone())
+                }
+                _ => None,
+            })
+            .collect();
+
+        Ok(deposits)
+    }
+
+    pub fn get_deposit_hash(&self, deposit_hashes: Vec<H256>) -> H256 {
+        if !deposit_hashes.is_empty() {
+            H256::from_slice(
+                [
+                    &(deposit_hashes.len() as u16).to_be_bytes(),
+                    &keccak(
+                        deposit_hashes
+                            .iter()
+                            .map(H256::as_bytes)
+                            .collect::<Vec<&[u8]>>()
+                            .concat(),
+                    )
+                    .as_bytes()[2..32],
+                ]
+                .concat()
+                .as_slice(),
+            )
+        } else {
+            H256::zero()
+        }
+    }
+    /// Prepare the state diff for the block.
+    pub fn prepare_state_diff(
+        &self,
+        block: &Block,
+        store: Store,
+        withdrawals: Vec<(H256, PrivilegedL2Transaction)>,
+        deposits: Vec<PrivilegedL2Transaction>,
+    ) -> Result<StateDiff, ProposerError> {
+        info!("Preparing state diff for block {}", block.header.number);
+
+        let mut state = evm_state(store.clone(), block.header.parent_hash);
+        execute_block(block, &mut state).map_err(ProposerError::from)?;
+        let account_updates = get_state_transitions(&mut state);
+
+        let mut modified_accounts = HashMap::new();
+        account_updates.iter().for_each(|account_update| {
+            modified_accounts.insert(
+                account_update.address,
+                AccountStateDiff {
+                    new_balance: account_update.info.clone().map(|info| info.balance),
+                    nonce_diff: account_update.info.clone().map(|info| info.nonce as u16),
+                    storage: account_update.added_storage.clone().into_iter().collect(),
+                    bytecode: account_update.code.clone(),
+                    bytecode_hash: None,
+                },
+            );
+        });
+
+        let state_diff = StateDiff {
+            modified_accounts,
+            version: StateDiff::default().version,
+            withdrawal_logs: withdrawals
+                .iter()
+                .map(|(hash, tx)| WithdrawalLog {
+                    address: match tx.to {
+                        TxKind::Call(address) => address,
+                        TxKind::Create => Address::zero(),
+                    },
+                    amount: tx.value,
+                    tx_hash: *hash,
+                })
+                .collect(),
+            deposit_logs: deposits
+                .iter()
+                .map(|tx| DepositLog {
+                    address: match tx.to {
+                        TxKind::Call(address) => address,
+                        TxKind::Create => Address::zero(),
+                    },
+                    amount: tx.value,
+                })
+                .collect(),
+        };
+
+        Ok(state_diff)
+    }
+
+    /// Generate the KZG commitment and proof for the blob. This commitment can then be used
+    /// to calculate the blob versioned hash, necessary for the EIP-4844 transaction.
+    pub fn prepare_blob_commitment(
+        &self,
+        state_diff: StateDiff,
+    ) -> Result<([u8; 48], [u8; 48]), ProposerError> {
+        let blob_data = state_diff.encode().map_err(ProposerError::from)?;
+
+        let blob = blob_from_bytes(blob_data).map_err(ProposerError::from)?;
+
+        let commitment = c_kzg::KzgCommitment::blob_to_kzg_commitment(&blob, self.kzg_settings)
+            .map_err(ProposerError::from)?;
+        let commitment_bytes =
+            Bytes48::from_bytes(commitment.as_slice()).map_err(ProposerError::from)?;
+        let proof =
+            c_kzg::KzgProof::compute_blob_kzg_proof(&blob, &commitment_bytes, self.kzg_settings)
+                .map_err(ProposerError::from)?;
+
+        let mut commitment_bytes = [0u8; 48];
+        commitment_bytes.copy_from_slice(commitment.as_slice());
+        let mut proof_bytes = [0u8; 48];
+        proof_bytes.copy_from_slice(proof.as_slice());
+
+        Ok((commitment_bytes, proof_bytes))
+    }
+
+    pub async fn send_commitment(
+        &self,
+        block_number: u64,
+        withdrawal_logs_merkle_root: H256,
+        deposit_logs_hash: H256,
+        commitment: [u8; 48],
+        proof: [u8; 48],
+        blob_data: Bytes,
+    ) -> Result<H256, ProposerError> {
+        info!("Sending commitment for block {block_number}");
+
+        let mut hasher = Sha256::new();
+        hasher.update(commitment);
+        let mut blob_versioned_hash = hasher.finalize();
+        blob_versioned_hash[0] = 0x01; // EIP-4844 versioning
+
+        let mut calldata = Vec::with_capacity(132);
+        calldata.extend(COMMIT_FUNCTION_SELECTOR);
+        let mut block_number_bytes = [0_u8; 32];
+        U256::from(block_number).to_big_endian(&mut block_number_bytes);
+        calldata.extend(block_number_bytes);
+        calldata.extend(blob_versioned_hash);
+        calldata.extend(withdrawal_logs_merkle_root.0);
+        calldata.extend(deposit_logs_hash.0);
+
+        let mut tx = EIP4844Transaction {
+            to: self.on_chain_proposer_address,
+            data: Bytes::from(calldata),
+            max_fee_per_gas: self.eth_client.get_gas_price().await?.as_u64(),
+            nonce: self.eth_client.get_nonce(self.l1_address).await?,
+            chain_id: self.eth_client.get_chain_id().await?.as_u64(),
+            blob_versioned_hashes: vec![H256::from_slice(&blob_versioned_hash)],
+            // TODO: blob_gas is too high, it should be set dynamically
+            max_fee_per_blob_gas: U256::from_dec_str("1000000").unwrap(),
+            // Should the max_priority_fee_per_gas be dynamic?
+            max_priority_fee_per_gas: 100u64,
+            ..Default::default()
+        };
+
+        let mut generic_tx = GenericTransaction::from(tx.clone());
+        generic_tx.from = self.l1_address;
+
+        tx.gas = self
+            .eth_client
+            .estimate_gas(generic_tx)
+            .await?
+            .saturating_add(TX_GAS_COST);
+
+        let mut buf = [0u8; BYTES_PER_BLOB];
+        buf.copy_from_slice(
+            blob_from_bytes(blob_data)
+                .map_err(ProposerError::from)?
+                .iter()
+                .as_slice(),
+        );
+
+        let mut wrapped_tx = WrappedEIP4844Transaction {
+            tx,
+            blobs_bundle: BlobsBundle {
+                blobs: vec![buf],
+                commitments: vec![commitment],
+                proofs: vec![proof],
+            },
+        };
+
+        let commit_tx_hash = self
+            .eth_client
+            .send_eip4844_transaction(&mut wrapped_tx, self.l1_private_key)
+            .await
+            .map_err(ProposerError::from)?;
+
+        info!("Commitment sent: {commit_tx_hash:#x}");
+
+        while self
+            .eth_client
+            .get_transaction_receipt(commit_tx_hash)
+            .await?
+            .is_none()
+        {
+            sleep(Duration::from_secs(1)).await;
+        }
+
+        Ok(commit_tx_hash)
+    }
+}
+
+pub async fn send_transaction_with_calldata(
+    eth_client: &EthClient,
+    l1_address: Address,
+    l1_private_key: SecretKey,
+    to: Address,
+    nonce: Option<u64>,
+    calldata: Bytes,
+) -> Result<H256, EthClientError> {
+    let mut tx = EIP1559Transaction {
+        to: TxKind::Call(to),
+        data: calldata,
+        max_fee_per_gas: eth_client.get_gas_price().await?.as_u64() * 2,
+        nonce: nonce.unwrap_or(eth_client.get_nonce(l1_address).await?),
+        chain_id: eth_client.get_chain_id().await?.as_u64(),
+        // Should the max_priority_fee_per_gas be dynamic?
+        max_priority_fee_per_gas: 10u64,
+        ..Default::default()
+    };
+
+    let mut generic_tx = GenericTransaction::from(tx.clone());
+    generic_tx.from = l1_address;
+
+    tx.gas_limit = eth_client
+        .estimate_gas(generic_tx)
+        .await?
+        .saturating_add(TX_GAS_COST);
+
+    eth_client
+        .send_eip1559_transaction(&mut tx, l1_private_key)
+        .await
+}
+
+async fn get_last_commited_block(
+    eth_client: &EthClient,
+    contract_address: Address,
+    overrides: Overrides,
+) -> Result<String, EthClientError> {
+    let selector = keccak(b"lastCommitedBlock()")
+        .as_bytes()
+        .get(..4)
+        .expect("Failed to get initialize selector")
+        .to_vec();
+
+    let mut calldata = Vec::new();
+    calldata.extend_from_slice(&selector);
+
+    eth_client
+        .call(contract_address, calldata.into(), overrides)
+        .await
+}
